+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - cmd/udns: add rrset query tool
+- DPRDataInfo.Type: add field to support API change
 
 ## [1.2.1] - 2016-06-13
 ### Fixed